renaming and addition to swagger docs
